Fixed middleware
